+++ test/test-cluster.js
@@ -1,6 +1,5 @@
 'use strict';
 const Redis = require('ioredis');
-const Promise = require('bluebird');
 const {makeTestSuite} = require('./_test-common');
 const port = Number(process.env.PORT) || 7001;
 const host = process.env.HOST || '127.0.0.1';
@@ -12,7 +11,6 @@ const redisClient = new Redis.Cluster([
 ], {
   scaleReads: 'all',
 });
-Promise.promisifyAll(redisClient);
 
 redisClient.on('error', (e) => {
   console.error(`Error connecting to Redis Cluster. Did you start a local server on ${port}?`);